Update lasagna.html
Added the image of the lasagna so
the visitors could see the end product
of following this recipe.

Added alt text so the visually impaired
could understand what the image
includes using their screen readers.

Added a title for hover-over text.

Adjusted the width of the image to
200 pixels so the image would not
take up a massive part of the page.