Update baked-spaghetti.html
Added a break-line to make the page
easier to read for the visitor.

Added an image of the final product
of the recipe so the visitors know what
they are in for.

Added alternative text for the visually
impaired to know what the image
contains, a title for hover-over text,
and adjusted the width so the image
doesn't take up a massive part of the
page.