Update fettucine-carbonara.html
Added a <br> element to make the
page easier to read for the user.

Added an image of the dish for the
visitors to know what they are about
to make.

Added an alternative text to the image
for the visually impaired to know what
the image contains using their screen
reader, a title for hover-over text
(attention to detail), and adjusted
the width of the image for the
image to not take up a massive part of
the page.